src/main/java/com/springapp/mvc/HelloController.java
package com.springapp.mvc;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Controller;
import org.springframework.ui.ModelMap;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

@Controller
@RequestMapping("/")
public class HelloController {
    private ModelMap modelMap;

    @Autowired
    public HelloController(ModelMap modelMap) {
        this.modelMap = modelMap;
    }

    @RequestMapping(method = RequestMethod.GET)
    public String printWelcome(ModelMap model) {
        model.addAttribute("message", "Interactive!");
        return "hello";
    }
}